pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-starter-python"
version = "1.0.0"
description = "Simple voice AI assistant built with LiveKit Agents for Python"
requires-python = ">=3.9"

dependencies = [
    "asyncio>=4.0.0",
    "dataclasses>=0.8",
    "google-genai>=1.32.0",
    "livekit-agents[aws,cartesia,deepgram,google,openai,resemble,sarvam,silero,turn-detector]~=1.2",
    "livekit-plugins-google>=1.2.6",
    "livekit-plugins-noise-cancellation~=0.2",
    "numpy>=2.0.2",
    "pathlib>=1.0.1",
    "pyaudio>=0.2.14",
    "pydub>=0.25.1",
    "python-docx>=1.2.0",
    "python-dotenv",
    "sounddevice>=0.5.2",
    "uuid>=1.30",
    "wavio>=0.0.9",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "A", "C4", "UP", "SIM", "RUF"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
